最終ポジションで Alpha タグが残らないように変更

+++ Assets/Sample.cs
@@ -38,7 +38,7 @@ public class Sample : MonoBehaviour
         Position.onValueChanged.AddListener(
             (val) =>
             {
-                RubyText.ForceTextPosition(RubyText.GetTextLength() * val);
+                RubyText.ForceTextPosition((int)((float)RubyText.GetTextLength() * val));
             }
         );
